🔨 [#053] - Split schedule-section.tsx into focused subcomponents and hooks 🗂️
- 🎨 Fixed S6759: added readonly to all props in EmployeeFormProps (employee-form.tsx)
- 🗂️ Created schedule-section-utils.ts with pure functions: buildSummaryLines, calcDayHours, formatHours, overrideDateLabel, computeDayRangeLabel
- 🧩 Created schedule-summary.tsx (ScheduleSummary component)
- 🔨 Created -use-dialog-animation.ts (useDialogAnimation hook + animCls helper)
- 🧩 Created schedule-dialog.tsx (ScheduleDialog modal/portal component)
- 🧩 Created create-schedule-form.tsx (CreateScheduleForm view, delegates to -use-create-schedule-inline)
- 🔨 Created -use-schedule-content.ts (useScheduleContent hook: all derived state + handlers extracted from ScheduleContent)
- 🧩 Created schedule-content.tsx (ScheduleContent thin JSX shell calling useScheduleContent)
- 🧩 Created schedule-config-rows.tsx (ReadRow, EditRow, DayLabel components)
- 🧩 Created override-scope-dialog.tsx (OverrideScopeDialog, delegates to use-override-scope-dialog)
- 🧩 Created weekly-calendar.tsx (WeeklyCalendar, WeekDayRow, OverrideListDialog)
- 🧩 Created schedule-section-state.tsx (EmptySchedule, ScheduleSkeleton)
- 🧹 Slimmed schedule-section.tsx from 1234 → 53 lines (ScheduleSection entry point only)
- 🔁 Re-exports utilities from schedule-section.tsx for backward compatibility with existing tests
