+++ code/webapp/src/components/employees/schedule-section-state.tsx
@@ -1,0 +1,28 @@
+import { CalendarDays } from 'lucide-react'
+
+export function EmptySchedule({ canCreate }: { readonly canCreate: boolean }) {
+  return (
+    <div className="flex flex-col items-center py-8 text-center">
+      <CalendarDays className="mb-2 h-8 w-8 text-muted-foreground/50" />
+      <p className="mb-1 text-sm font-medium">Sin horario activo</p>
+      <p className="mb-4 text-xs text-muted-foreground">Este empleado no tiene un horario vigente configurado.</p>
+      {!canCreate && <p className="text-xs text-muted-foreground">El empleado no tiene un período laboral activo.</p>}
+    </div>
+  )
+}
+
+export function ScheduleSkeleton() {
+  return (
+    <div className="space-y-3">
+      <div className="h-4 w-40 animate-pulse rounded bg-muted" />
+      {Array.from({ length: 5 }, (_, i) => (
+        <div key={i} className="flex gap-4">
+          <div className="h-4 w-16 animate-pulse rounded bg-muted" />
+          <div className="h-4 w-10 animate-pulse rounded bg-muted" />
+          <div className="h-4 w-10 animate-pulse rounded bg-muted" />
+          <div className="h-4 w-10 animate-pulse rounded bg-muted" />
+        </div>
+      ))}
+    </div>
+  )
+}
