+++ code/webapp/src/components/employees/schedule-summary.tsx
@@ -1,0 +1,17 @@
+import type { EmployeeSchedule } from '@/types/schedule'
+import { buildSummaryLines } from './schedule-section-utils'
+
+export function ScheduleSummary({ schedule }: { readonly schedule: EmployeeSchedule }) {
+  const lines = buildSummaryLines(schedule.days)
+  if (lines.length === 0) return null
+
+  return (
+    <div className="mt-1.5 space-y-0.5 pl-0.5">
+      {lines.map((line) => (
+        <div key={line.icon} className="flex items-center gap-1.5 text-xs text-muted-foreground">
+          <span>{line.text}</span>
+        </div>
+      ))}
+    </div>
+  )
+}
